requirements.txt
langchain
langchain-ollama

PyYAML~=6.0.2